# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Receipt Raven API",
    "description": "API for processing and managing receipts",
    "version": "1.0.0"
  },
  "components": {
    "schemas": {
      "receipt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "payee": {
            "type": "string"
          },
          "expenseAccount": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "totalSum": {
            "type": "number"
          },
          "status": {
            "type": "string"
          },
          "filepath": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "payee",
          "date",
          "totalSum",
          "status",
          "filepath"
        ],
        "title": "receipt"
      },
      "receiptItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "receiptId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "price": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "receiptId",
          "name",
          "price",
          "category"
        ],
        "title": "receiptItem"
      },
      "receiptSubmission": {
        "type": "object",
        "properties": {
          "expenseAccount": {
            "type": "string"
          },
          "payee": {
            "type": "string"
          },
          "date": {
            "type": "string"
          },
          "totalSum": {
            "type": "number"
          },
          "items": {
            "type": "array",
            "minItems": 1,
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number"
                },
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "category": {
                  "type": "string"
                }
              },
              "required": [
                "name",
                "price",
                "category"
              ]
            }
          }
        },
        "required": [
          "expenseAccount",
          "payee",
          "date",
          "totalSum",
          "items"
        ],
        "title": "receiptSubmission"
      }
    }
  },
  "paths": {
    "/api/upload": {
      "post": {
        "tags": [
          "upload"
        ],
        "description": "Upload a receipt image for processing",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "filename": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "success",
                    "filename"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/accounts": {
      "get": {
        "tags": [
          "ledger"
        ],
        "description": "Get all accounts from the ledger",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "type",
            "required": false,
            "description": "Filter by account type (e.g., \"Expenses\")"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/categories": {
      "get": {
        "tags": [
          "ledger"
        ],
        "description": "Get expense categories from the ledger",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "account_mapping": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "name",
                      "account_mapping"
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/payees": {
      "get": {
        "tags": [
          "ledger"
        ],
        "description": "Get list of payees (shops, vendors) from the ledger",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "payees": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      }
                    }
                  },
                  "required": [
                    "payees"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/submit-receipt/{receiptId}": {
      "post": {
        "tags": [
          "ledger"
        ],
        "description": "Submit a receipt to the ledger as a transaction",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "receiptId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error",
                    "message"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts": {
      "get": {
        "tags": [
          "receipts"
        ],
        "description": "Get receipts",
        "responses": {
          "200": {
            "description": "List of receipts",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of receipts",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/receipt"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/{receiptId}": {
      "post": {
        "tags": [
          "receipts"
        ],
        "description": "Save receipt and its items to the ledger",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/receiptSubmission"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "receiptId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "success"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/receipts/{receiptId}/items": {
      "get": {
        "tags": [
          "receipts"
        ],
        "description": "Get receipt items",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "receiptId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "List of receipt items in a receipt",
            "content": {
              "application/json": {
                "schema": {
                  "description": "List of receipt items in a receipt",
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/receiptItem"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "description": "Health check endpoint",
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "ok"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Development server"
    }
  ],
  "tags": [
    {
      "name": "health",
      "description": "Health check endpoints"
    },
    {
      "name": "upload",
      "description": "Receipt upload endpoints"
    },
    {
      "name": "ledger",
      "description": "Ledger integration endpoints"
    }
  ]
}