# backend/src/modules/ledger/routes.ts
import { FastifyPluginAsync } from 'fastify';

const ledgerRoutes: FastifyPluginAsync = async fastify => {
  /**
   * GET /api/ledger/accounts
   * Get all accounts from the ledger
   * Query params:
   *  - type: optional filter by account type (e.g., "Expenses")
   */
  fastify.get<{
    Querystring: { type?: string };
  }>('/ledger/accounts', {
    schema: {
      tags: ['ledger'],
      description: 'Get all accounts from the ledger',
      querystring: {
        type: 'object',
        properties: {
          type: {
            type: 'string',
            description: 'Filter by account type (e.g., "Expenses")',
          },
        },
      },
      response: {
        200: {
          type: 'array',
          items: {
            type: 'string',
          },
        },
        500: {
          type: 'object',
          properties: {
            error: { type: 'string' },
            message: { type: 'string' },
          },
          required: ['error', 'message'],
        },
      },
    },
    handler: async (request, reply) => {
      try {
        const { type } = request.query;
        const accounts = await fastify.ledgerService.getAccounts(type);
        return accounts.map(a => a.name);
      } catch (error) {
        fastify.log.error({ error }, 'Failed to fetch accounts from ledger');
        return reply.code(500).send({
          error: 'Failed to fetch accounts',
          message: error instanceof Error ? error.message : 'Unknown error',
        });
      }
    },
  });

  /**
   * GET /api/ledger/categories
   * Get expense categories from the ledger
   */
  fastify.get('/ledger/categories', {
    schema: {
      tags: ['ledger'],
      description: 'Get expense categories from the ledger',
      response: {
        200: {
          type: 'array',
          items: {
            type: 'object',
            properties: {
              name: { type: 'string' },
              account_mapping: { type: 'string' },
            },
            required: ['name', 'account_mapping'],
          },
        },
        500: {
          type: 'object',
          properties: {
            error: { type: 'string' },
            message: { type: 'string' },
          },
          required: ['error', 'message'],
        },
      },
    },
    handler: async (request, reply) => {
      try {
        const categories = await fastify.ledgerService.getCategories();
        console.log(categories);
        return { categories };
      } catch (error) {
        fastify.log.error({ error }, 'Failed to fetch categories from ledger');
        return reply.code(500).send({
          error: 'Failed to fetch categories',
          message: error instanceof Error ? error.message : 'Unknown error',
        });
      }
    },
  });

  /**
   * GET /api/ledger/payees
   * Get list of payees (shops, vendors) from the ledger
   */
  fastify.get('/ledger/payees', {
    schema: {
      tags: ['ledger'],
      description: 'Get list of payees (shops, vendors) from the ledger',
      response: {
        200: {
          type: 'object',
          properties: {
            payees: {
              type: 'array',
              items: { type: 'string' },
            },
          },
          required: ['payees'],
        },
        500: {
          type: 'object',
          properties: {
            error: { type: 'string' },
            message: { type: 'string' },
          },
          required: ['error', 'message'],
        },
      },
    },
    handler: async (request, reply) => {
      try {
        const payees = await fastify.ledgerService.getPayees();
        return { payees };
      } catch (error) {
        fastify.log.error({ error }, 'Failed to fetch payees from ledger');
        return reply.code(500).send({
          error: 'Failed to fetch payees',
          message: error instanceof Error ? error.message : 'Unknown error',
        });
      }
    },
  });

  /**
   * POST /api/ledger/submit-receipt/:receiptId
   * Submit a receipt to the ledger as a transaction
   * Body should contain source account and any additional transaction details
   */
  fastify.post<{
    Params: { receiptId: string };
    Body: {
      sourceAccount: string;
      currency?: string;
      itemAccountMappings?: Record<string, string>; // Map item names to expense accounts
    };
  }>('/ledger/submit-receipt/:receiptId', {
    schema: {
      tags: ['ledger'],
      description: 'Submit a receipt to the ledger as a transaction',
      params: {
        type: 'object',
        properties: {
          receiptId: { type: 'string' },
        },
        required: ['receiptId'],
      },
      response: {
        200: {
          type: 'object',
          properties: {
            success: { type: 'boolean' },
            message: { type: 'string' },
          },
        },
        400: {
          type: 'object',
          properties: {
            error: { type: 'string' },
          },
          required: ['error'],
        },
        404: {
          type: 'object',
          properties: {
            error: { type: 'string' },
          },
          required: ['error'],
        },
        500: {
          type: 'object',
          properties: {
            error: { type: 'string' },
            message: { type: 'string' },
          },
          required: ['error', 'message'],
        },
      },
    },
    handler: async (request, reply) => {
      try {
        const receiptId = parseInt(request.params.receiptId);

        // TODO: remove receipt db dependency from ledger module and pass all data here
        // via http request!
        // Fetch receipt data from database
        const receipt =
          await fastify.receiptRepository.getReceiptById(receiptId);
        if (!receipt) {
          return reply.code(404).send({
            error: 'Receipt not found',
          });
        }

        if (!receipt.expense_account) {
          return reply.code(400).send({
            error: 'Receipt is missing the expense account',
          });
        }

        const receiptItems =
          await fastify.receiptRepository.getReceiptItems(receiptId);

        // Submit to ledger service
        const result = await fastify.ledgerService.submitReceiptTransaction({
          receipt_id: receiptId,
          payee: receipt.payee,
          date: receipt.receipt_date,
          expense_account: receipt.expense_account,
          items: receiptItems,
          total: receipt.total_sum,
        });

        return result;
      } catch (error) {
        fastify.log.error({ error }, 'Failed to submit receipt to ledger');
        return reply.code(500).send({
          error: 'Failed to submit receipt',
          message: error instanceof Error ? error.message : 'Unknown error',
        });
      }
    },
  });
};

export default ledgerRoutes;